moved the util method to component too
